Move eTLDPlusOne to brave_wallet_utils
