+++ browser/brave_wallet/BUILD.gn
@@ -56,7 +56,7 @@ source_set("brave_wallet_delegate") {
     "//brave/browser/brave_wallet",
     "//brave/components/brave_wallet/browser",
     "//brave/components/brave_wallet/browser:constants",
-    "//brave/components/brave_wallet/browser:ethereum_permission_utils",
+    "//brave/components/brave_wallet/browser:utils",
     "//brave/components/brave_wallet/common:mojom",
     "//brave/components/resources:strings_grit",
     "//components/content_settings/core/common",
