+++ components/brave_wallet/browser/brave_wallet_utils_unittest.cc
@@ -1036,4 +1036,14 @@ TEST(BraveWalletUtilsUnitTest, GetCurrentChainId) {
             mojom::kFilecoinMainnet);
 }
 
+TEST(BraveWalletUtilsUnitTest, eTLDPlusOne) {
+  EXPECT_EQ("", eTLDPlusOne(GURL()));
+  EXPECT_EQ("brave.com", eTLDPlusOne(GURL("https://blog.brave.com")));
+  EXPECT_EQ("brave.com", eTLDPlusOne(GURL("https://...brave.com")));
+  EXPECT_EQ("brave.com", eTLDPlusOne(GURL("https://a.b.c.d.brave.com/1")));
+  EXPECT_EQ("brave.github.io",
+            eTLDPlusOne(GURL("https://a.b.brave.github.io/example")));
+  EXPECT_EQ("", eTLDPlusOne(GURL("https://github.io")));
+}
+
 }  // namespace brave_wallet
